fix： parseSizeToBytes
